+++ examples/desugar.rs
@@ -1,10 +1,10 @@
 use
 {
-	thespis         :: { *                         } ,
-	thespis_impl    :: { *                         } ,
-	async_executors :: { AsyncStd, SpawnHandleExt  } ,
-	std             :: { error::Error              } ,
-	futures         :: { channel::mpsc             } ,
+	thespis         :: { *                        } ,
+	thespis_impl    :: { *                        } ,
+	async_executors :: { AsyncStd, SpawnHandleExt } ,
+	std             :: { error::Error             } ,
+	futures         :: { channel::mpsc            } ,
 };
 
 
@@ -34,10 +34,11 @@ impl Handler< Hello > for MyActor
 //
 async fn main() -> Result< (), Box<dyn Error> >
 {
-	let (tx, rx)  = mpsc::channel( 5 );
-	let mb        = Mailbox::new( Some("HelloWorld".into()), Box::new(rx) );
-	let mut addr  = Addr::new( mb.id(), mb.name(), Box::new( tx.sink_map_err( |e| Box::new(e) as SinkError ) ) );
-	let actor     = MyActor;
+	let (tx, rx)  = mpsc::channel( 5 )                                          ;
+	let tx        = Box::new( tx.sink_map_err( |e| Box::new(e) as SinkError ) ) ;
+	let mb        = Mailbox::new( Some("HelloWorld"), Box::new(rx) )            ;
+	let mut addr  = mb.addr( tx )                                               ;
+	let actor     = MyActor                                                     ;
 
 	let handle = AsyncStd.spawn_handle( mb.start( actor ) )?;