+++ examples/supervisor.rs
@@ -56,7 +56,7 @@ struct Supervisor
 //
 struct Supervise<A: Actor>
 {
-	mailbox : Option< JoinHandle<Option<Mailbox<A>>> > ,
+	mailbox : Option< JoinHandle<MailboxEnd<A>> > ,
 
 	// A closure that knows how to instantiate the supervised actor.
 	//
@@ -99,9 +99,9 @@ impl<A: Actor + Send> Handler< Supervise<A> > for Supervisor
 			// This is where the magic happens. Every time the handle resolves, we spawn again
 			// and replace it with a new handle.
 			//
-			// When this returns None, it means the actor has stopped naturally and we don't respawn it.
+			// When this returns MailboxEnd::Actor, it means the actor has stopped naturally and we don't respawn it.
 			//
-			while let Some(mb) = mb_handle.await
+			while let MailboxEnd::Mailbox(mb) = mb_handle.await
 			{
 				mb_handle = mb.start_handle( (actor.create)(), &AsyncStd ).unwrap();
 			}