+++ tests/errors.rs
@@ -3,6 +3,9 @@
 // ✔ let thread of mailbox crash and verify mailbox closed error
 // ✔ let thread of mailbox crash and verify mailbox closed before response error
 //
+#![ cfg(not( target_arch = "wasm32" )) ]
+
+
 mod common;
 
 use
@@ -54,7 +57,7 @@ async fn test_mb_closed() -> Result<(), DynError >
 		Err(e) => match e
 		{
 			ThesErr::MailboxClosed{..} => {}
-			_                          => assert!( false, "Wrong error returned: {:?}", e ),
+			_                          => panic!( "Wrong error returned: {:?}", e ),
 		}
 	}
 
@@ -103,7 +106,7 @@ async fn test_mb_closed_before_response() -> Result<(), DynError >
 		Err(e) => match e
 		{
 			ThesErr::ActorStoppedBeforeResponse{..} => {}
-			_ => assert!( false, "Wrong error returned: {:?}", e ),
+			_ => panic!( "Wrong error returned: {:?}", e ),
 		}
 	}